[zsh] better ~/tmp creation
Former-commit-id: 0e70566c1add7cb6657b08f32a06a31c5d743316 [formerly 5797f8c25bf8797e32379772410af7d9b85ea585] [formerly 636b435ca18b8ad34e7af0f5fcd35c9b113670e9 [formerly 6409362d6140e51e6e9c522b09119d779f3e4a41]]
Former-commit-id: 8c739fa1aec0a17677fdf2c5907673e135a71292 [formerly 56bba1402182501c7d2bcbd44bbf776dae22ee6b]
Former-commit-id: 2d76d9f12c3ebc9aaa9e28aea4d542d93aa1de0b

+++ .zshenv
@@ -1,3 +1,2 @@
-if [[ ! -L ~/tmp ]]; then
-    rm -f ~/tmp && tmp_loc=$(mktemp -d) && ln -fs ${tmp_loc} ~/tmp
-fi
+[[ $(readlink -e ~/tmp) == "" ]] && rm -f ~/tmp
+[[ ! -L ~/tmp ]] && { rm -f ~/tmp && tmp_loc=$(mktemp -d) && ln -fs ${tmp_loc} ~/tmp }
